I reduce page complexity for WAYSPACE mvp.
