--- pages/create/index.js
@@ -1,7 +1,0 @@
-import CreatePage from '@components/CreatePage/CreatePage'
-
-const Create = () => {
-  return <CreatePage />
-}
-
-export default Create
